tool2-essay grader/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.9
pydantic-settings==2.5.2
sqlalchemy==2.0.23
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1